+++ aiootp/_permutations/fast_affine_xor_chain.py
@@ -68,7 +68,7 @@ class FastAffineXORChain(AffineXORChain):
                 permutation_config_id=config_id,
                 key_types=AffineXORChainConfig._FAST_CHAIN_KEY_TYPES,
             )
-            for config_id in [*range(1, 33), 64, 128, 192, 256]
+            for config_id in [*range(1, 33), 48, 64, 128, 192, 256]
         },
         config_type=AffineXORChainConfig,
     )