+++ CHANGES.rst
@@ -20,6 +20,7 @@ Minor Changes
 ^^^^^^^^^^^^^
 
 - fix: capture ``KeyError`` & raise ``AttributeError`` instead [#8]
+- feat: add 48-byte default permutation config objects
 - build: use dynamic table to specify the package readme
 - build: specify the coverage htmlcov directory
 - build: remove 'command_line' so coverage can run specific files